+++ package.json
@@ -15,15 +15,15 @@
   },
   "devDependencies": {
     "@biomejs/biome": "^2.1.3",
-    "@types/node": "^22.13.9",
-    "@types/node-fetch": "^2.6.12",
-    "cross-env": "^7.0.3",
-    "dotenv": "^16.4.7",
-    "jest": "^29.7.0",
+    "@types/node": "^24.2.0",
+    "@types/node-fetch": "^2.6.13",
+    "cross-env": "^10.0.0",
+    "dotenv": "^17.2.1",
+    "jest": "^30.0.5",
     "source-map-support": "^0.5.21",
-    "tsup": "^8.4.0",
-    "turbo": "^2.4.4",
-    "zx": "^8.3.2"
+    "tsup": "^8.5.0",
+    "turbo": "^2.5.5",
+    "zx": "^8.8.0"
   },
   "packageManager": "pnpm@9.15.4",
   "workspaces": [
@@ -32,9 +32,9 @@
     "examples/astro-example"
   ],
   "dependencies": {
-    "lit": "^3.2.1",
-    "typedoc": "^0.27.9",
-    "wrangler": "^3.112.0"
+    "lit": "^3.3.1",
+    "typedoc": "^0.28.9",
+    "wrangler": "^4.28.0"
   },
   "version": "1.0.0"
 }
