+++ packages/up-provider/typedoc.json
@@ -1,5 +1,14 @@
 {
   "$schema": "https://typedoc.org/schema.json",
   "entryPoints": ["./src/index.ts"],
-  "includeVersion": true
+  "includeVersion": true,
+  "externalSymbolLinkMappings": {
+    "lit": {
+      "css": "#",
+      "PropertyDeclarations": "#"
+    },
+    "@lit/reactive-element": {
+      "PropertyDeclaration": "#"
+    }
+  }
 }
